Moved thumb generation into its own Python file
The code was getting too long and my scroll wheel got tired.
